Update GCCS-Metrics_DataSourcesCleaning_and_Join.R.R
Deleted unnecessary code
Fixed errors
Using MCPD standards
Always INSTCODE instead of the institute names

Genesys dataset:
No need to add a column with the institute name, there is the INSTCODE and that is enough. Dropped from the code. 

No need to convert insitute code in  DUPLSITE to institute names, using the insitutte codes is the best solution to record sites of safety duplication. 

Country codes:
some of the country codes entered manually are not correct, I deleted the corresponding code lines: XAE is CIMMYT and XAM is IRRI as origin of the material, it should not changed to a (wrong) country, 

BGCI dataset issues identified

The country in the BGCI dataset is not the country of origin but the country of the institute holding the accession. Therefore, the code for this part was deleted. 

BGCI dataset: institute code to be added, we need to ask BGCI if they have a conversion table for their garden ID to INSTCODE. Otherwise, a conversion table has to be done manually. 

WIEWS: 
INSTCODE should always be used as an identifier for an institute (it avoids errors generated from different versions of the same name for the same institute)